requirements.txt
flask
mysql-connector-python
configparser
flask_bcrypt
flask_sqlalchemy